refactor: extract plugin marketplace components
