+++ tenvy-server/src/routes/(app)/plugins/+page.svelte
@@ -1,73 +1,42 @@
 <script lang="ts">
 	import { cn } from '$lib/utils.js';
-	import { Badge } from '$lib/components/ui/badge/index.js';
-	import { Button } from '$lib/components/ui/button/index.js';
+	import MarketplaceGrid from '$lib/components/plugins/MarketplaceGrid.svelte';
+	import PluginCard from '$lib/components/plugins/PluginCard.svelte';
 	import {
-		Card,
-		CardContent,
-		CardDescription,
-		CardFooter,
-		CardHeader,
-		CardTitle
-	} from '$lib/components/ui/card/index.js';
-	import { Input } from '$lib/components/ui/input/index.js';
-	import { Separator } from '$lib/components/ui/separator/index.js';
-	import { Switch } from '$lib/components/ui/switch/index.js';
+		marketplaceStatusStyles,
+		distributionNotice,
+		formatSignatureTime,
+		signatureBadge
+	} from '$lib/components/plugins/utils.js';
+        import { Badge } from '$lib/components/ui/badge/index.js';
+        import { Button } from '$lib/components/ui/button/index.js';
         import {
-                pluginCategories,
-                pluginCategoryLabels,
-                pluginDeliveryModeLabels,
-                pluginStatusLabels,
-                pluginStatusStyles,
-                formatRelativeTime,
-                type Plugin,
-                type PluginCategory,
-                type PluginDeliveryMode,
-                type PluginStatus,
-                type PluginUpdatePayload
-        } from '$lib/data/plugin-view.js';
-	import type { PluginManifest } from '../../../../../shared/types/plugin-manifest.js';
-	import type { UserRole } from '$lib/server/auth.js';
+                Card,
+                CardContent,
+                CardDescription,
+                CardFooter,
+                CardHeader,
+                CardTitle
+        } from '$lib/components/ui/card/index.js';
+        import { Input } from '$lib/components/ui/input/index.js';
+        import { Separator } from '$lib/components/ui/separator/index.js';
+        import { Switch } from '$lib/components/ui/switch/index.js';
 	import {
-		Check,
-		Download,
-		Info,
-		PackageSearch,
-		RefreshCcw,
-		Search,
-		ShieldAlert,
-		SlidersHorizontal,
-		Wifi,
-		ShieldCheck,
-		GitFork
-	} from '@lucide/svelte';
-
-	type MarketplaceStatus = 'pending' | 'approved' | 'rejected';
-
-        type MarketplaceListing = {
-                id: string;
-                name: string;
-                summary: string | null;
-                repositoryUrl: string;
-                version: string;
-                pricingTier: string;
-                status: MarketplaceStatus;
-                manifest: PluginManifest;
-                submittedBy: string | null;
-                reviewerId: string | null;
-                signature: Plugin['signature'];
-        };
-
-	type MarketplaceEntitlement = {
-		id: string;
-		listingId: string;
-		tenantId: string;
-		seats: number;
-		status: string;
-		listing: MarketplaceListing;
-	};
-
-        type AuthenticatedUser = { id: string; role: UserRole } | null;
+		pluginCategories,
+		pluginCategoryLabels,
+		pluginStatusLabels,
+		pluginStatusStyles,
+		type Plugin,
+		type PluginCategory,
+		type PluginStatus,
+		type PluginUpdatePayload
+	} from '$lib/data/plugin-view.js';
+	import type {
+		AuthenticatedUser,
+		MarketplaceEntitlement,
+		MarketplaceListing
+	} from '$lib/components/plugins/types.js';
+	import { Check, Info, RefreshCcw, Search, SlidersHorizontal } from '@lucide/svelte';
 
 	let {
 		data
@@ -103,7 +72,7 @@
 	let marketplaceEntitlements = $state<MarketplaceEntitlement[]>(
 		data.entitlements.map((entitlement) => ({ ...entitlement }))
 	);
-        const currentUser = $state<AuthenticatedUser>(data.user ?? null);
+	const currentUser = $state<AuthenticatedUser>(data.user ?? null);
 	let searchTerm = $state('');
 	let statusFilter = $state<'all' | PluginStatus>('all');
 	let categoryFilter = $state<'all' | PluginCategory>('all');
@@ -156,80 +125,12 @@
 		}
 	}
 
-	const distributionModes: PluginDeliveryMode[] = ['manual', 'automatic'];
-
-	function distributionNotice(plugin: Plugin): string {
-		if (!plugin.enabled) return 'Plugin currently disabled';
-
-		const notes = [`Default: ${pluginDeliveryModeLabels[plugin.distribution.defaultMode]}`];
-
-		if (!plugin.distribution.allowManualPush) {
-			notes.push('manual pushes blocked');
-		}
-
-		if (!plugin.distribution.allowAutoSync) {
-			notes.push('auto-sync paused');
-		}
-
-		return notes.join(' · ');
-	}
-
 	function resetFilters() {
 		searchTerm = '';
 		statusFilter = 'all';
 		categoryFilter = 'all';
 		autoUpdateOnly = false;
 	}
-
-        function statusSeverity(status: PluginStatus) {
-                switch (status) {
-                        case 'error':
-                                return 'text-red-500';
-                        case 'update':
-                                return 'text-amber-500';
-                        default:
-                                return 'text-muted-foreground';
-                }
-        }
-
-        function signatureBadge(signature: Plugin['signature']) {
-                switch (signature.status) {
-                        case 'trusted':
-                                return {
-                                        label: 'Signature trusted',
-                                        icon: ShieldCheck,
-                                        class: 'border-emerald-500/40 text-emerald-500'
-                                } as const;
-                        case 'unsigned':
-                                return {
-                                        label: 'Unsigned',
-                                        icon: ShieldAlert,
-                                        class: 'border-amber-500/40 text-amber-500'
-                                } as const;
-                        case 'untrusted':
-                                return {
-                                        label: 'Untrusted signature',
-                                        icon: ShieldAlert,
-                                        class: 'border-amber-500/40 text-amber-500'
-                                } as const;
-                        case 'invalid':
-                        default:
-                                return {
-                                        label: 'Signature invalid',
-                                        icon: ShieldAlert,
-                                        class: 'border-red-500/40 text-red-500'
-                                } as const;
-                }
-        }
-
-        function formatSignatureTime(value: string | null | undefined): string {
-                if (!value) return 'never';
-                const parsed = new Date(value);
-                if (Number.isNaN(parsed.valueOf())) {
-                        return value;
-                }
-                return formatRelativeTime(parsed);
-        }
 
 	const normalizedSearch = $derived(searchTerm.trim().toLowerCase());
 
@@ -272,23 +173,11 @@
 			autoUpdateOnly
 	);
 
-	const listingStatusStyles: Record<MarketplaceStatus, string> = {
-		approved: 'border border-emerald-500/40 bg-emerald-500/10 text-emerald-500',
-		pending: 'border border-amber-500/40 bg-amber-500/10 text-amber-500',
-		rejected: 'border border-red-500/40 bg-red-500/10 text-red-500'
-	};
+	const canPurchase = $derived(currentUser?.role === 'admin' || currentUser?.role === 'operator');
 
-	function isEntitled(listingId: string): boolean {
-		return marketplaceEntitlements.some((entry) => entry.listingId === listingId);
-	}
-
-        const canPurchase = $derived(
-                currentUser?.role === 'admin' || currentUser?.role === 'operator'
-        );
-
-        const canSubmitMarketplace = $derived(
-                currentUser?.role === 'admin' || currentUser?.role === 'developer'
-        );
+	const canSubmitMarketplace = $derived(
+		currentUser?.role === 'admin' || currentUser?.role === 'developer'
+	);
 
 	async function purchaseListing(listing: MarketplaceListing) {
 		if (!canPurchase) {
@@ -318,133 +207,16 @@
 </script>
 
 <section class="space-y-6">
-	<Card class="border-border/60">
-		<CardHeader class="flex flex-col gap-2 sm:flex-row sm:items-center sm:justify-between">
-			<div class="space-y-1">
-				<CardTitle>Marketplace</CardTitle>
-				<CardDescription>
-					Deploy community plugins backed by signed releases from public repositories.
-				</CardDescription>
-			</div>
-			<Badge variant="secondary" class="gap-2 text-xs">
-				<ShieldCheck class="h-4 w-4" />
-				{marketplaceListings.length} listing{marketplaceListings.length === 1 ? '' : 's'}
-			</Badge>
-		</CardHeader>
-		<CardContent>
-			{#if marketplaceListings.length === 0}
-				<p class="text-sm text-muted-foreground">
-					No marketplace submissions yet. Developers can publish plugins once approved by an
-					administrator.
-				</p>
-			{:else}
-				<div class="grid gap-4 lg:grid-cols-2 xl:grid-cols-3">
-					{#each marketplaceListings as listing (listing.id)}
-						<div
-							class="flex flex-col justify-between rounded-lg border border-border bg-card p-4 shadow-sm"
-						>
-							<div class="space-y-3">
-                                                                {@const listingSignature = signatureBadge(listing.signature)}
-                                                                <div class="flex items-start justify-between gap-3">
-                                                                        <div class="space-y-1">
-                                                                                <h3 class="text-base leading-tight font-semibold">{listing.name}</h3>
-                                                                                <p class="text-xs tracking-wide text-muted-foreground uppercase">
-                                                                                        Version {listing.version} · {listing.pricingTier}
-                                                                                </p>
-                                                                        </div>
-                                                                        <div class="flex flex-col items-end gap-2">
-                                                                                <Badge class={listingStatusStyles[listing.status]}>{listing.status}</Badge>
-                                                                                <Badge
-                                                                                        variant="outline"
-                                                                                        class={cn(
-                                                                                                'flex items-center gap-1 rounded-full border px-2 py-1 text-[10px] font-semibold uppercase tracking-wide',
-                                                                                                listingSignature.class
-                                                                                        )}
-                                                                                >
-                                                                                        <svelte:component
-                                                                                                this={listingSignature.icon}
-                                                                                                class="h-3 w-3"
-                                                                                        />
-                                                                                        {listingSignature.label}
-                                                                                </Badge>
-                                                                        </div>
-                                                                </div>
-                                                                <p class="text-sm leading-relaxed text-muted-foreground">
-                                                                        {listing.summary ?? listing.manifest.description ?? 'No description provided.'}
-                                                                </p>
-                                                                <div class="flex flex-col gap-2 text-xs text-muted-foreground">
-									<div class="flex items-center gap-2">
-										<GitFork class="h-3.5 w-3.5" />
-										<a
-											href={listing.repositoryUrl}
-											rel="noreferrer"
-											target="_blank"
-											class="truncate underline decoration-dotted hover:text-foreground"
-										>
-											{listing.repositoryUrl}
-										</a>
-                                                                        </div>
-                                                                        <div class="flex items-center gap-2">
-                                                                                <ShieldCheck class="h-3.5 w-3.5" />
-                                                                                <span>{listing.manifest.license.spdxId}</span>
-                                                                        </div>
-                                                                        <div class="flex items-center gap-2">
-                                                                                <svelte:component
-                                                                                        this={listingSignature.icon}
-                                                                                        class="h-3.5 w-3.5"
-                                                                                />
-                                                                                <span>
-                                                                                        {listing.signature.signer ?? listing.signature.publicKey ??
-                                                                                        listingSignature.label}
-                                                                                </span>
-                                                                        </div>
-                                                                        <div class="flex items-center gap-2">
-                                                                                <Info class="h-3.5 w-3.5" />
-                                                                                <span>
-                                                                                        Checked {formatSignatureTime(listing.signature.checkedAt ?? null)}
-                                                                                </span>
-                                                                        </div>
-                                                                        {#if listing.signature.error}
-                                                                                <p class="text-xs text-red-500">
-                                                                                        Signature error: {listing.signature.error}
-                                                                                </p>
-                                                                        {/if}
-                                                                </div>
-                                                        </div>
-							<div class="mt-4 flex items-center justify-between">
-								<span class="text-xs text-muted-foreground">
-									{listing.manifest.capabilities?.length ?? 0} capability{(listing.manifest
-										.capabilities?.length ?? 0) === 1
-										? ''
-										: 'ies'}
-								</span>
-								<Button
-									size="sm"
-									variant={isEntitled(listing.id) ? 'outline' : 'default'}
-									disabled={!canPurchase || isEntitled(listing.id) || listing.status !== 'approved'}
-									onclick={() => purchaseListing(listing)}
-								>
-									{#if isEntitled(listing.id)}
-										Entitled
-									{:else if listing.status !== 'approved'}
-										Awaiting review
-									{:else}
-										Purchase
-									{/if}
-								</Button>
-							</div>
-						</div>
-					{/each}
-				</div>
-			{/if}
-		</CardContent>
-		{#if canSubmitMarketplace}
-			<CardFooter class="text-xs text-muted-foreground">
-				Developers can submit new plugins via the controller API after uploading signed release
-				assets to GitHub.
-			</CardFooter>
-		{/if}
-	</Card>
+	<MarketplaceGrid
+		listings={marketplaceListings}
+		entitlements={marketplaceEntitlements}
+		{canPurchase}
+		{canSubmitMarketplace}
+		{purchaseListing}
+		{signatureBadge}
+		{formatSignatureTime}
+		statusStyles={marketplaceStatusStyles}
+	/>
 	<Card class="border-border/60">
 		<CardHeader class="space-y-4">
 			<div class="flex flex-col gap-4 lg:flex-row lg:items-center lg:justify-between">
@@ -558,280 +330,7 @@
 	{#if filteredPlugins.length > 0}
 		<div class="grid gap-4">
 			{#each filteredPlugins as plugin (plugin.id)}
-				<Card
-					class={cn(
-						'border-border/60 transition',
-						plugin.status === 'error' && 'border-red-500/40',
-						plugin.status === 'update' && 'border-amber-500/40',
-						!plugin.enabled && 'opacity-90'
-					)}
-				>
-					<CardHeader class="flex flex-col gap-4 lg:flex-row lg:items-start lg:justify-between">
-						<div class="space-y-2">
-							<div class="flex flex-wrap items-center gap-3">
-								<CardTitle class="text-base leading-tight font-semibold">{plugin.name}</CardTitle>
-								<Badge
-									variant="outline"
-									class="px-2.5 py-1 text-xs font-medium text-muted-foreground"
-								>
-									v{plugin.version}
-								</Badge>
-                                                                <Badge
-                                                                        variant="outline"
-                                                                        class={cn('px-2.5 py-1 text-xs font-medium', pluginStatusStyles[plugin.status])}
-                                                                >
-                                                                        {pluginStatusLabels[plugin.status]}
-                                                                </Badge>
-                                                                {@const sigBadge = signatureBadge(plugin.signature)}
-                                                                <Badge
-                                                                        variant="outline"
-                                                                        class={cn('px-2.5 py-1 text-xs font-medium flex items-center gap-1', sigBadge.class)}
-                                                                >
-                                                                        <svelte:component this={sigBadge.icon} class="h-3.5 w-3.5" />
-                                                                        {sigBadge.label}
-                                                                </Badge>
-                                                        </div>
-                                                        <CardDescription class="max-w-2xl text-sm text-muted-foreground"
-                                                                >{plugin.description}</CardDescription
-                                                        >
-							<div class="flex flex-wrap items-center gap-2">
-								{#each plugin.capabilities as capability (capability)}
-									<Badge variant="secondary" class="bg-muted text-muted-foreground">
-										{capability}
-									</Badge>
-								{/each}
-								{#if plugin.requiredModules.length > 0}
-									<span
-										class="text-[10px] font-semibold tracking-wide text-muted-foreground uppercase"
-									>
-										Requires
-									</span>
-									{#each plugin.requiredModules as module (module.id)}
-										<Badge
-											variant="secondary"
-											class="border border-border/60 bg-background/60 text-foreground"
-										>
-											{module.title}
-										</Badge>
-									{/each}
-								{/if}
-							</div>
-						</div>
-						<div class="flex flex-col gap-4 text-sm text-muted-foreground">
-							<div class="flex items-center gap-2">
-								<Info class={cn('h-4 w-4', statusSeverity(plugin.status))} />
-								<span class="font-medium text-foreground"
-									>{pluginCategoryLabels[plugin.category]}</span
-								>
-							</div>
-							<div class="grid gap-1">
-								<span
-									>Maintainer: <strong class="font-medium text-foreground">{plugin.author}</strong
-									></span
-								>
-								<span>Last deployed {plugin.lastDeployed}</span>
-								<span>Health check {plugin.lastChecked}</span>
-							</div>
-						</div>
-					</CardHeader>
-					<CardContent class="grid gap-4 lg:grid-cols-2">
-                                                <div class="grid gap-4 text-sm text-muted-foreground md:grid-cols-2 xl:grid-cols-3">
-                                                        <div class="space-y-1 rounded-md border border-border/60 px-3 py-2">
-                                                                <span class="text-xs tracking-wide uppercase">Signature</span>
-                                                                {@const sig = signatureBadge(plugin.signature)}
-                                                                <p class="flex items-center gap-2 text-sm font-semibold text-foreground">
-                                                                        <svelte:component this={sig.icon} class="h-4 w-4" />
-                                                                        {sig.label}
-                                                                </p>
-                                                                {#if plugin.signature.error}
-                                                                        <p class="text-xs text-muted-foreground">{plugin.signature.error}</p>
-                                                                {:else if plugin.signature.signer}
-                                                                        <p class="text-xs text-muted-foreground">Signer: {plugin.signature.signer}</p>
-                                                                {/if}
-                                                                <p class="text-xs text-muted-foreground">
-                                                                        Checked {formatSignatureTime(plugin.signature.checkedAt)}
-                                                                </p>
-                                                        </div>
-                                                        <div class="space-y-1 rounded-md border border-border/60 px-3 py-2">
-                                                                <span class="text-xs tracking-wide uppercase">Installations</span>
-                                                                <p class="text-lg font-semibold text-foreground">{plugin.installations}</p>
-                                                        </div>
-							<div class="space-y-1 rounded-md border border-border/60 px-3 py-2">
-								<span class="text-xs tracking-wide uppercase">Package size</span>
-								<p class="text-lg font-semibold text-foreground">{plugin.size}</p>
-							</div>
-							<div class="space-y-1 rounded-md border border-border/60 px-3 py-2">
-								<span class="text-xs tracking-wide uppercase">Status</span>
-								<p class="text-lg font-semibold text-foreground">
-									{pluginStatusLabels[plugin.status]}
-								</p>
-							</div>
-							<div class="space-y-1 rounded-md border border-border/60 px-3 py-2">
-								<div class="flex items-center justify-between">
-									<span class="text-xs tracking-wide uppercase">Manual deployments</span>
-									<Download class="h-4 w-4 text-muted-foreground" />
-								</div>
-								<p class="text-lg font-semibold text-foreground">
-									{plugin.distribution.manualTargets}
-								</p>
-								<p class="text-xs text-muted-foreground">
-									Last push {plugin.distribution.lastManualPush}
-								</p>
-							</div>
-							<div class="space-y-1 rounded-md border border-border/60 px-3 py-2">
-								<div class="flex items-center justify-between">
-									<span class="text-xs tracking-wide uppercase">Auto enrollments</span>
-									<Wifi class="h-4 w-4 text-muted-foreground" />
-								</div>
-								<p class="text-lg font-semibold text-foreground">
-									{plugin.distribution.autoTargets}
-								</p>
-								<p class="text-xs text-muted-foreground">
-									Last sync {plugin.distribution.lastAutoSync}
-								</p>
-							</div>
-							<div class="space-y-1 rounded-md border border-border/60 px-3 py-2">
-								<div class="flex items-center justify-between">
-									<span class="text-xs tracking-wide uppercase">Package artifact</span>
-									<PackageSearch class="h-4 w-4 text-muted-foreground" />
-								</div>
-								<p class="font-medium break-words text-foreground">{plugin.artifact}</p>
-								<p class="text-xs text-muted-foreground">
-									Default: {pluginDeliveryModeLabels[plugin.distribution.defaultMode]}
-								</p>
-							</div>
-						</div>
-						<div class="flex flex-col gap-3">
-							<div
-								class="flex items-center justify-between rounded-md border border-border/60 px-3 py-2"
-							>
-								<div class="space-y-1">
-									<p class="text-sm leading-tight font-medium">Plugin enabled</p>
-									<p class="text-xs leading-tight text-muted-foreground">
-										Controls whether the module can run on assigned clients.
-									</p>
-								</div>
-								<Switch
-									checked={plugin.enabled}
-									aria-label={`Toggle ${plugin.name}`}
-									onCheckedChange={(value) => {
-										const nextStatus = value
-											? plugin.status === 'disabled'
-												? 'active'
-												: plugin.status
-											: 'disabled';
-
-										void updatePlugin(plugin.id, {
-											enabled: value,
-											status: nextStatus
-										});
-									}}
-								/>
-							</div>
-							<div
-								class="flex items-center justify-between rounded-md border border-border/60 px-3 py-2"
-							>
-								<div class="space-y-1">
-									<p class="text-sm leading-tight font-medium">Automatic updates</p>
-									<p class="text-xs leading-tight text-muted-foreground">
-										When enabled, new builds roll out without manual approval.
-									</p>
-								</div>
-								<Switch
-									checked={plugin.autoUpdate}
-									aria-label={`Toggle auto update for ${plugin.name}`}
-									onCheckedChange={(value) => void updatePlugin(plugin.id, { autoUpdate: value })}
-								/>
-							</div>
-							<div class="space-y-3 rounded-md border border-border/60 px-3 py-2">
-								<div class="space-y-1">
-									<p class="text-sm leading-tight font-medium">Delivery mode</p>
-									<p class="text-xs leading-tight text-muted-foreground">
-										Choose how the plugin is distributed to agents and clients.
-									</p>
-								</div>
-								<div class="flex flex-wrap gap-2">
-									{#each distributionModes as mode (mode)}
-										<Button
-											type="button"
-											size="sm"
-											variant={plugin.distribution.defaultMode === mode ? 'default' : 'outline'}
-											disabled={!plugin.enabled}
-											aria-pressed={plugin.distribution.defaultMode === mode}
-											onclick={() =>
-												void updatePlugin(plugin.id, {
-													distribution: {
-														defaultMode: mode
-													}
-												})}
-										>
-											{pluginDeliveryModeLabels[mode]}
-										</Button>
-									{/each}
-								</div>
-								<div class="grid gap-3 sm:grid-cols-2">
-									<div
-										class="flex items-center justify-between rounded-md border border-dashed border-border/60 px-3 py-2"
-									>
-										<div class="min-w-0 space-y-1">
-											<p class="text-sm leading-tight font-medium">Allow manual downloads</p>
-											<p class="text-xs leading-tight text-muted-foreground">
-												Permit operators to push the package to specific targets.
-											</p>
-										</div>
-										<Switch
-											checked={plugin.distribution.allowManualPush}
-											disabled={!plugin.enabled}
-											aria-label={`Toggle manual downloads for ${plugin.name}`}
-											onCheckedChange={(value) =>
-												void updatePlugin(plugin.id, {
-													distribution: {
-														allowManualPush: value
-													}
-												})}
-										/>
-									</div>
-									<div
-										class="flex items-center justify-between rounded-md border border-dashed border-border/60 px-3 py-2"
-									>
-										<div class="min-w-0 space-y-1">
-											<p class="text-sm leading-tight font-medium">Allow auto-sync</p>
-											<p class="text-xs leading-tight text-muted-foreground">
-												Auto-download the plugin whenever an agent connects.
-											</p>
-										</div>
-										<Switch
-											checked={plugin.distribution.allowAutoSync}
-											disabled={!plugin.enabled}
-											aria-label={`Toggle auto sync for ${plugin.name}`}
-											onCheckedChange={(value) =>
-												void updatePlugin(plugin.id, {
-													distribution: {
-														allowAutoSync: value
-													}
-												})}
-										/>
-									</div>
-								</div>
-							</div>
-						</div>
-					</CardContent>
-					<CardFooter class="flex flex-wrap items-center justify-between gap-3">
-						<div
-							class="flex items-center gap-2 text-xs tracking-wide text-muted-foreground uppercase"
-						>
-							<ShieldAlert class="h-4 w-4" />
-							{distributionNotice(plugin)}
-						</div>
-						<div class="flex items-center gap-2">
-							<Button type="button" variant="outline" size="sm" class="gap-2">
-								<RefreshCcw class="h-4 w-4" />
-								Check for updates
-							</Button>
-							<Button type="button" size="sm" variant="ghost">Open details</Button>
-						</div>
-					</CardFooter>
-				</Card>
+				<PluginCard {plugin} {updatePlugin} {distributionNotice} />
 			{/each}
 		</div>
 	{:else}
